Fold measure plan rules into one file
Merge the former p0 rule set into defaultMeasurePlanRules
and drop tier labels from checks.json.
